refactor: Define the service and shuvi scope
