+++ packages/service/src/config/index.ts
@@ -2,9 +2,4 @@ import { ShuviConfig } from './configTypes';
 
 export type { ShuviConfig };
 
-export type OptionsKeyMap = Record<
-  string,
-  string | ((config: any, optionValue: any) => void)
->;
-
-export { normalizeConfig, loadConfig, defineConfig } from './config';
+export { loadConfig } from './config';
