+++ packages/shuvi/src/cmds/build.ts
@@ -1,9 +1,9 @@
 import program from 'commander';
 import * as path from 'path';
 import logger from '@shuvi/utils/lib/logger';
-import { OptionsKeyMap } from '@shuvi/service';
 import { build } from '../tasks/build';
-import { getPackageInfo, getProjectDir, getConfigFromCli } from '../utils';
+import { getPackageInfo, getProjectDir } from '../utils';
+import { OptionsKeyMap, getConfigFromCli } from '../config';
 
 const pkgInfo = getPackageInfo();
