+++ packages/shuvi/src/config/index.ts
@@ -1,0 +1,8 @@
+export type OptionsKeyMap = Record<
+  string,
+  string | ((config: any, optionValue: any) => void)
+>;
+
+export { normalizeConfig, defineConfig } from './config';
+
+export { getConfigFromCli } from './utils';
