+++ packages/shuvi/src/index.ts
@@ -1,7 +1,7 @@
 import '@shuvi/platform-web/shuvi-type-extensions-node';
 
 export { ShuviMode, ShuviPhase } from './types';
-export { ShuviConfig, loadConfig, defineConfig } from '@shuvi/service';
+export { defineConfig } from './config';
 export type { PresetFunction } from '@shuvi/service';
 export {
   // api
@@ -11,5 +11,8 @@ export {
   PHASE_PRODUCTION_BUILD,
   PHASE_PRODUCTION_SERVER,
   PHASE_DEVELOPMENT_SERVER,
-  PHASE_INSPECT_WEBPACK
+  PHASE_INSPECT_WEBPACK,
+  // config
+  ShuviConfig,
+  loadConfig
 } from '@shuvi/service';
