+++ packages/shuvi/src/shuvi.ts
@@ -1,12 +1,7 @@
-import {
-  IPlatform,
-  getApi,
-  Api,
-  ShuviConfig,
-  normalizeConfig
-} from '@shuvi/service';
+import { IPlatform, getApi, Api, ShuviConfig } from '@shuvi/service';
 import platformWeb from '@shuvi/platform-web';
 import { ShuviMode, ShuviPhase } from './types';
+import { normalizeConfig } from './config';
 
 export function getPlatform(): IPlatform {
   return platformWeb({ framework: 'react' });
